Allow `{async_}resource_ref` to be constructible from a pointer to a resource
Currently all APIs take a pointer to e.g. a `rmm::device_memory_resource`

With the current design that we need a reference to construct a `{async_}resource_ref` we are forced to rewrite all those APIs.

Instead, allow construction from a pointer to resource and allow simple gradual transition between APIs