Remove having two install rules for <project>-header-search.cmake
The install(DIRECTORY would generate one rule with an incorrect
source file. This became an issue when a git clone and cmake
configuration was executed within 1sec. CMake uses mtime
to determine if a file is out of date, and in these cases
the second proper install rule wouldn't run.