Add NVRTC_SKIP_KERNEL_RUN tag to compile, but skip running NVRTC test (#434)
Some kernels using TMA need some host code to run as well. In addition,
they need the device symbols to be accessible from host code. This turns
out to be quite difficult to achieve. Therefore, since the device code
is already executed in the non-nvrtc tests, it is reasonable to allow a
test to skip running device code when NVRTC is enabled. This still tests
whether the code actually compiles.