[libcu++] Fix undefined behavior in atomics to automatic storage
The current implementation of atomic operations is unsound.
It issues generic PTX atomic instructions even if the address
falls in the local memory address space, causing well-formed
CUDA C++ programs to exhibit PTX undefined behavior.

Since this only impact objects with automatic storage, the
impact is not very widespread, but it does impact beginners
trying to learn libcu++ atomic operations, and it also impacts
most of the examples in our documentation which use automatic
storage for simplicity.

This change tests whether the address of an atomic operation is
in local memory using `__isLocal`, and when that is the case, it
uses weak memory operations instead. This is sound because CUDA C++
does not allow sharing the address of automatic variables across
threads. If that ever changes, this would need to be updated.

Unfortunately, nvidia compilers from toolkits older than 12.3 have
a bug that miscompiles programs that use `__isLocal`, like our
workaround here. Instead, we use PTX `isspace` instruction to
perform the detection.