Remove c2h::copy_if as it doesn't build on MSVC.
The filtered type lists are now hardcoded as a WAR.

Also removed the warning suppression for the unused var in catch2, as
local windows builds did not respect the pragma. Now we just hardcode
whether FP types are available for each key pass and disable the
relevant tests as needed.