Move ptx.md out of extended API
Previously, ptx.md was part of the extended API, but in the left hand
side bar was listed as its own section.

Now, it has been moved out of the extended API, and moved to a separate
spot in the left-hand side bar, below Extended API, and above Releases,
and Contributing.