Match_any testing single bit by fusing into single LOP3 instruction
I also observed that implementation of the match_any function could also be optimized a bit. Presently, it employs "setp.eq.u32 p, %0, %2;" instead of "setp.ne.u32 p, %0, 0;". While functionally this is equivalent when testing a single bit, adopting != 0 could enable a fusion of the 'and' and 'testp' operations into a single LOP3 instruction in area where ALU intensity is high.