Fix include of <thrust/random.h> with NVC++
PR #1320 broke compilation with NVC++ when `<thrust/random.h>` was the
first header to be included.  This code that was there previously
```
    defined(_CCCL_COMPILER_CLANG) || \
    defined(_CCCL_COMPILER_ICC)
```
does not include `<stdint.h>` when NVC++ is the compiler, but the file
uses types defined in `<stdint.h>`, such as:
```
typedef ::int8_t   int8_t;
typedef ::int16_t  int16_t;
```

Fix this problem by using the same `#if` condition for including
`<stdint.h>` as for using the ``::intN_t` types.