[BACKPORT]: Fix ptx usage to account for PTX ISA availability (#1359) (#1421)
* Fix ptx usage to account for PTX ISA availability (#1359)

Currently we only guard those instructions based on the available architecture. However, it is also valid to compile with an old toolkit for a new machine.

Consequently we need to strengthen our checks against available PTX ISA

* Do not use VLAs in `cp_async_bulk_tensor_*` tests

VLAs are a compiler extension and are correctly errored out by some compilers. As we always know the exact size of the array anyway just swtich to a `cuda::std::array`

Fixes nvbug4476664

* Use proper shared memory size

Authored-by: Allard Hendriksen <ahendriksen@nvidia.com>

* Fix incorrect linker issue

* Ensure runfail tests do not fail without execution

* Ensure that __cccl_ptx_isa properly guards feature flags