Fix thrust::optional<T&>::emplace()
Where optional<T> inherits optional<T>::construct via a series
of classes, optional<T&> does not. This means that
optional<T&>::emplace() was broken and called into a member function
that did not exist.

This replaces the functionality to make optional<T&>::emplace() change
the stored reference to the new one. Note that it does _not_ emplace
the referee, as this would lead to questionable behavior when the
optional holds nullopt.

This was revealed by a change in LLVM, see
https://github.com/llvm/llvm-project/pull/90152 and
https://github.com/ROCm/rocThrust/issues/404.