Prioritize CI jobs with dependencies.
Currently, all standalone jobs tend to start long before the two stage producers.
This severely impacts CI turnaround time as the longest job chains usually are the last to start.

GHA doesn't provide any way to prioritize jobs other than full `needs:` dependencies, which are massively overkill for this usecase.

This patch attempts to solve this by assuming that all jobs within a matrix strategy will be queued in the order specified.
This eliminates the standalone jobs, moving them to two-stage jobs without consumers.
The two-stage jobs with consumers are sorted in front of the consumerless jobs, hopefully encouraging them to start first.