Drop all internal implementations of exceptions (#1806) (#1839)
* Drop all internal implementations of exceptions

It turns out that our implementations of exceptions are a bad idea.

Conventionally a lot of user will check for e.g. `::std::exception` and catch that or anything. But we are actually defining our own exceptions which a user would have to explicitly also catch.

At the same time, exceptions are only available on host anyway, so the benefit of using them is ... slim.

This makes it so that we always inherit from `std::exception`

Co-authored-by: Wesley Maxey <71408887+wmaxey@users.noreply.github.com>