Make `cuda::std::tuple` trivially copyable
This is similar to the situation with `cuda::std::pair`
We have a lot of users that rely on types being trivially copyable, so that they can utilize memcpy and friends.

Previously, `cuda::std::tuple` did not satisfy this because it needs to handle reference types.

Given that we already specialize `__tuple_leaf` depending on whether the class is empty or not, we can simply
add a third specialization that handles the trivially copyable types and one that synthesizes assignment.