[PoC]: Implement `cuda::experimental::uninitialized_buffer` (#1831)
* Drop `cuda::get_property` CPO

It serves no purpose as it only ever forwards via ADL and also breaks older nvcc

* Ensure that we test memory resources

* Implement `cuda::uninitialized_buffer`

`cuda::uninitialized_buffer` provides an allocation of `N` elements of type `T` utilitzing a `cuda::mr::resource` to allocate the storage.

`cuda::uninitialized_buffer` takes care of alignment and deallocation of the storage. The user is required to ensure that the lifetime of the memory resource exceeds the lifetime of the buffer.