Implement P0843 `inplace_vector` (#1936)
* Implement P0843 `inplace_vector`

This implements `inplace_vector` a resizable container with a fixed capacity that stores its elements in a local array.

Due to the fact that everything is local we are more or less save regarding host device issues, as long as users do not pass around references of it.

The exception guarantees are not 100% clear yet, so I implemented them on a best effort basis. We might need to revisit what we guarantee in case of an exception.