Fix `zip_function` together with `proclaim_return_type`
When using `zip_function` with `proclaim_return_type` it wants to deduce the return type from the invoke expression inside of `proclaim_return_type`
which is not possible because device lambdas do not have a return type.

This seems to be a nvcc bug, but one we can work around with.

Fixes #2344