add IWYU `export` pragma on config headers (#2352)
When working in the devcontainers, the clangd/clang-format integration causes extra #includes to be added automatically. They get added because clang's "include what you use" tool (IWYU) doesn't consider transitive includes when determining what headers need to be included.

This PR adds the export IWYU pragma to the #includes in CCCL's config headers so that a #include of <cuda/__cccl_config> or <cuda/std/detail/__config> will behave as if all of the transitive headers were included directly. This should be enough to keep IWYU from automatically adding headers it shouldn't.