Propagate compiler flags down to libcu++ LIT tests
In order to run libcu++ tests with clang lit has to use correct compiler settings.
It currently appears to rely on clang automatically finding and using
the default CUDA installation, but that's not necessarily the CUDA that
the build itself may have been configured to use.

Clang builds may also require building them with the host libc++, instead of the default libstdc++.
That requires propagation of the top-level CFLAGS specified by the user.