Require resources to always provide at least one execution space property
Currently we implicitly assumed that any resource that had no execution space property was host accessible.

However, that is not a good design, as it provides a source of surprise and numerous challenges with proper type matching down the road.

So rather than implicitly assuming that something is host accessible, we require the user to always provide at least one execution space property.