Fix `common_type` specialization for extended floating point types
The machinery we had in place was not really suited to specialize `common_type` because it would take precendence over the actual implementation of `common_type`

In that case, we only specialized `common_type<__half, __half>` but not `common_type<__half, __half&>` and so on.

This shows how brittle the whole thing is and that it is not extensible.

Rather than putting another bandaid over it, add a proper 5th step in the common_type detection that properly treats combinations of an extended floating point type with an arithmetic type.

Allowing arithmetic types it necessary to keep machinery like `pow(__half, 2)` working.

Fixes [BUG]: `is_common_type`  trait is broken when mixing rvalue references #2419