Fix value range overflows in tests
`random_integers<fp-type>()` may generate values not representable as an integer type,
which results in undefined behavior when we attempt to cast them to integer type later on.

To avoid that, generate random numbers using the integer type we'll cast to later on,
so we're guaranteed to have representable values.