Make get_global_resource host/device
The function is called from a host/device constructors and may be invoked from the device code.

As it happens neither NVCC nor clang diagnose such an invalid call.
https://github.com/llvm/llvm-project/issues/118415

As the result, in unoptimized builds we end up with ptxas failing with
an unresolved reference to this function because it is never generated
during GPU-side compilation.
https://github.com/NVIDIA/cccl/pull/2813#issuecomment-2477099734