Alway include cuda_runtime_api with stream_ref
`cuda::stream_ref` does not work without at least including cuda_runtime_api.h

We previously guarded this by using a cuda compiler but it should be possible to use this with just a c++ compiler.

This makes it so and also adds stream_ref to the host compiler tests.

The one caveat is that we need to ensure that we properly provide the cuda include path so the compiler finds cuda_runtime_api.h