Revert back to cub::Traits::CATEGORY
Users specialize cub::Traits, adding new types (e.g. custom half or bfloat16 types) which are classified as PRIMITIVE and FLOATING_POINT. The traits in libcu++ will not classify those types as such without the user specialization them additionally. Since this is a breaking change, let's revert it for CCCL 2.8.