Add test_reduce_api.py::test_reduce_struct_type_minmax
This adds an example of using struct that holds minimum and maximum values
to compute smallest and largest values of an array using a single call
cub::DeviceReduce

It also demonstrates a use of TransformedIterator to map input array
values as v -> MinMax(v,v) to make reduction binary operator commutative,
to render data-parallel algorithm applicable. Since this is done in
registers, each element of input array is only accessed once.