Add segmented_reduce python api (#3906)
* Add algorithms.segmented_reduce Python API

Also avoid recomputing cccl_value of init in both segmented_reduce
and in reduce

* Change to input_array fixture

1. Include np.complex64
2. Device output size in a variable and reuse it to avoid repeated
   occurrances of literal values
3. Generate real/imag values for complex arrays in a single
   call to sampling function for efficiency
4. Change range of generated integral arrays based on the signness
   of the integral data type. For unsigned types we continue to
   sample in interval [0, 10), for signed we sample from [-5, 5].

* Corrected docstring of segmented_reduce function

* Add initial tests for segmented_reduce

* Improve readability of test_segmented_reduce_api example

* TransformIteratorKind need not override __eq__/__hash__ methods of the base

Additionally, changed the __hash__ of IteratorKind to mix the hash of
its value with hash of self.__class__.

* Add AdvancedIterator(it, offset=1) function

This is used to advance a given iterator `it` the `offset` steps without
running into multiple definitions of the advance/derefence methods.

* Add example for summing rows of a matrix using segmented_reduce

* Implement IteratorBase.__add__(self, offset : int) using make_advanced_iterator

* Use end_offsets = start_offsets + 1

This calls IteratorBase.__add__ to produce an iterator whose state
is advanced by 1, but which shares the same advance/dereference methods.

* Add a test for segmented_reduce on gpu_struct

* Change hash of transform iterator to mix its kind

* Rename variable n to sample_size

Also make generation of complex array in test_reduce.py more
efficient by genering real and imaginary components in a single
call to np.random.random instead of using two calls.

* Remove __hash__ and __eq__ special methods from some iterator classes

These were only defined for TransformIterator and AdvancedIterator classes,
but not for other classes.

Implemented review suggestion to type type(self) instead of self.__class__

* Tweak test_scan_array_input to avoid integer overflows during host accumulation

For short range data types we take a small slice of the input array to
avoid running into the overflow problem. This works because input_array
fixture samples from uniform discrete distribution with small upper range (8),
hence using 31 uint8 elements can run up to 31 * 7  = 217 ( < 255) and fits
in the type.

* Add cccl.set_cccl_iterator_state utility function and use in segmented_reduce.py

* Introduce _bindings.call_build utility

This finds compute capability and include paths and appends them
to the algorithm-specific arguments. Used the utility in segmented_reduce.

* Make call_build take *args, **kwargs