add a visitation interface to the senders of ustdex
this commit adds a `visit` free function to ustdex. it is callable with
a visitor, a sender, and an extra blob of user data. it expands one
level of the sender tree. the visitor is responsible for recursing.

`visit` uses the fact that senders from the standard algorithms can be
used to initialize a structured binding. this commit adds a trait
`structured_binding_size` that is specialized for all the sender types.
that trait is used to dispatch to a visitation function of the correct
arity. for now the arity is capped at 8 (tag+data+6 children) but that
can be increased. eventually the limit will go away with C++26's
variadic structured binding support.

this commit also uses the preprocessor to unroll tuples of up to 8
elements, thereby instantiating O(1) class templates instead of O(N)+1.
the `when_all` sender uses an unrolled tuple to store its child senders
so that it satisfies the requirements for structured bindings without
going through the `std::tuple` interface (which is expensive at compile
time).