Turn `cuda::std::iter_swap` into a CPO to avoid ADL fiasko
If there is a type that brings in both namespaces `::std` and `::cuda::std`, then we will currently error out because `iter_swap` is ambiguous.

Unfortunately, we cannot use the usual techniques to break the ambigouity in favor of `std::iter_swap` because the algorithm accepts two different iterator types and deduction fails if we try to wrap one in `type_identity_t`

Given the choice between breaking valid code that pulls in namespace `::std` and valid code that use ADL for `::cuda::std::iter_swap` we opt to break the code relying on ADL because that can actually be fixed by adding qualifiers.