Add classes to enable build caching in c.parallel test suite
1. `result_wrapper_t` is owning reference to CCCL build result based
on shared pointer. It must be copy/move assignable/constructable
to enable use as value type in std::unordered_map.

The class is templated on CCCL build result struct type, and
callable to clean-up resources allocated for build result.

2. `build_cache_t` in simple cache based on std::unordered_map.

The class is templated on key type (std::string) and value type,
an instance of result_wrapper_t.

3. `fixture` is a singleton to lazily create build cache at
  its first use. Fixture wraps the cache in std::optional
  to match use pattern in algorithms.

fixture allows for Tag template parameter to be able to create
multiple instances of singleton classes. We usually want cache
for a test family where algorithm compile the same problem and
execute it for different problem sizes. Caching then allows to
reuse the build artifact.

4. KeyBuilder is a struct with static methods to facilitate
   creation of test-variant identifier to use as cache key.

Cache keys are crafted per test.