Change validation of large num_segments tests in Python
Validation is now performed by constructing an iterator yielding
expected result. The allocating fixed size uint8 buffer and running
binary transform to compare actual to expected and write result
of comparison to the validation buffer. We then use
`cp.all(validation.view(cp.bool_))` to validate this chunk, and
move to the next.