Fix bug re UDTs and items_per_thread > 1.
The lowering error Numba CUDA was reporting was simply due to it not
knowing how to lower the `initial_value` `Complex()` that was being used
in the tests.

Adding in a `lower_constant` to `Complex` was sufficient to fix the
issue.  There are now no restrictions regarding user-defined types and
items per thread greater than 1.