Fix cuda.coop limitation preventing user-defined types when items_per_thread > 1 in block scan module. (#4756)
* Fix bug re UDTs and items_per_thread > 1.

The lowering error Numba CUDA was reporting was simply due to it not
knowing how to lower the `initial_value` `Complex()` that was being used
in the tests.

Adding in a `lower_constant` to `Complex` was sufficient to fix the
issue.  There are now no restrictions regarding user-defined types and
items per thread greater than 1.

* Change the default items_per_thread from 1 to 4.

Now that there are no restrictions to usage when items_per_thread is
greater than 1 (i.e. we fixed the user-defined types issue), it makes
sense to use a more optimal default that should yield better performance
out-of-the-box for most use cases.

* PR Feedback: make `items_per_thread` a mandatory argument.

Per @gevtushenko's suggestion.

* Fix test missed in first pass.