Run CCCL infra tests when example projects may have changed. (#4816)
We've had some frequent nightly failures because the CCCL Infra tests weren't running for changes to core projects that are used in the example CMake projects.
These tests are very fast and cheap, and running them for libcudacxx, cub, thrust, and cudax changes will plug this coverage gap in the PR workflow.

* Ensure that only direct dependencies affect project dirty state.