Speedup runtime of c/parallel/test/test_radix_sort.cpp (#4848)
For every one of 4 test key types, the test was running 4 combinations
where is_descending was true/false, and is_overwrite_okay was
independently was true/false.

This resulted in 16 distinct compilations. This change pairs it down
to replacing direct product of iteration spaces with zip of them,
reducing it to 4 distinct compilations.

This way we still exercise each of distinct choices of boolean
build parameters but at a fewer number of builds