Make `cuda::stream_ref` an env for itself
We want to move to environment based APIs.
However, sometimes it would be nice to direclty pass a `cuda::stream_ref` around.

That does not work because `cuda::stream_ref` is not a query for `cuda::get_stream_t`

So turn it into one so that `cuda::stream_ref` can return itself on that query