Generate a version from git/JSON for packages (#4889)
Adding some versioning information to the wheels packages will enable the solver to install the latest pre-release or post-release packages, closes #4900. This uses `setuptools_scm` to determine a version using `git` by default, but then overrides that in CI with a custom version string that has a prefix and optional suffix if not on a clean tag.

It uses the version from `cccl-version.json` by default, and the commit count from the last tag. In the future it could move to just the last tag and the commit count once we have tags on `main` that more closely match what is desired.