Add dependency on numba-cuda
Dependency on numba-cuda is necessary to ensure discoverability
of NVVM library installed in Python layout. numba-cuda relies
on pynvjitlink-cu12 to accomplish this (to the best of my understanding),
but treats it as an optional dependency.

For this reason, we specify run-time dependency on pynvjitlink-cu12
in cuda-cccl metadata. A comment is added to this effect.