get the `starts_on` algorithm working with the stream scheduler (#5092)
fixing the `starts_on` algorithm for the stream scheduler involved
first switching from early customization (in the sender algorithms)
to lazy customization (in `connect`). with this commit, senders are
adapted to the stream scheduler just-in-time, when they are being
connected to a receiver.

this has the advantage that the receiver can provide the stream
the work will be run on.