Unify `_Byte_holder` class for all the hasher's
Reduce code duplication by unifiying `_Byte_holder` class. A new template parameter `_UseTailBlock`
is added to distinguish how tailing bytes are treated. The template parameter is set to true for
xxhash, as it treats tail bytes to be blocks which are incomplete chunks. Whereas, for murmurhash based
hasher it is set to false, as murmurhash treats all the tail bytes as just bytes instead of blocks.