Fix some issues withe the calculation of `__fp_one`
The problem occured with floating point types without an implicit bit like `long double` which however takes a different code path.

The ther issue was that we did not properly cast the integer of the bias to the storage type, which could again could lead to issues with types that have more mantissa bits than `int`.

However, currently the only supported types that satisfy this are builtin types, which again take a different code path.

Still worth fixing though