Drop all usages of `_CCCL_TRAIT`
This was a cludge to make C++17 code more efficient thatn C++14 was able to do. But now we dropped C++14, so we can always just use the inline variables.

I used the following two regexes to replace most of the occurences:
`_CCCL_TRAIT\(([a-zA-Z_:0-9]*), ([a-zA-Z_:0-9<>, .&]*)\)`
`_CCCL_TRAIT\(([a-zA-Z_:0-9]*), ([a-zA-Z_:0-9<>, .&]*), ([a-zA-Z_:0-9<>, .&]*)\)`

The remaining 100 cases were quickly done by hand.