Fix style of functions' titles
As noted in gh-5694, rendered CUB documentation
uses code-block styling for generated function
pages, resulting in jarred style in TOC for CUB
documentation.

This commit uses plain text instead, consistent
with how it is done for class/struct/typedef/enum
qualifiers.

The code to inject this header in the content of
generated RST file is moved few lines later to
when qualified name is computed.

The result is that TOC now uses uniform style.