Remove problematic new build symlink (#5761)
* Remove duplicate job entries.

* Fix build issue and restore previous behavior.

The addition of 'local' to our build infix when using will break workflows that are directly configuring presets outside of a container by moving the build directory.

We can restore the previous behavior by removing the symlink that gets recursive when CCCL_BUILD_INFIX is unset.

Now builds outside of devcontainers will build to the same location whether using presets or scripts.