Redacted some comments in util_type CUDA header file for clarity (#5868)
* Reworded comments explaining member types of UnitWord<T> struct

Used math to convey the objective more clearly.

* Move comment about specialization being recursion base to improve formatting

Move the comment for the base case specialization to the end of struct Log2
definition, in line with where it is planed in the recursive case, so that
clang formatter would form the value of VALUE enum as a single-line conditional
expression.