Allow CI to run on forks with sccache enabled. (#5882)
* Allow CI to run on forks with sccache enabled.

Adds a minimal 'pull_request_fork' workflow that can be launched on github public runners.
This workflow will automatically run when a pull request is opened or updated *on a fork* of CCCL. It will not run CI automatically on every push, and the copy-pr-bot logic remains in place on the main repo.

The python wheel builder was updated to remove the temporary docker images after it is done with them, as they were exhausting the available storage on github runners.

* Avoid needing read:enterprise scope until sccache-dist is ready.