Improve interoperability of `cuda` iterators with thrust and std (#5929)
* Also check for `iterator_concept` with the C++17 iterator checks

* Impove our iterator interoperability with iterator_traits

* Fix template argument order of `cuda::transform_*_iterator`

I swapped the order of the arguments, because that is the common order for standard algorithms.

However, that wil break our transition from `thrust::transform_iterator` becuase users would have to swap arguments actively and not simply change the namespace

* Relax constraints on the cuda iterators

Thrust iterators often do not satisfy all constraints  to qualify as `random_access_iterator` but they provide alll the operations.

Rather than potentially slowing down a lot of code and break users, relax the constraints on the cuda iterators, so that we can interoperate with the thrust ones.