Get Windows c.parallel build working. (#5924)
* Fix CMakeLists.txt for jit_templates on Windows.

Make sure we call the correct MSVC pre-processor correctly, and, use a
more robust file concat technique that works with the CMake that ships
with Visual Studio 2022 (3.31.6-msvc6).

* Fix OffsetT type to be `long long` on Windows only for transform.cu.

`long` on Linux is 64-bit, but 32-bit on Windows.  Surprisingly, using
`long long` on both (i.e. for Linux as well, without the #ifdef guards)
results in a Linux compilation error (the static assert trips).

* Fix "missing return statement at end of non-void function" on MSVC.

  T:\src\cccl\c\parallel\src\transform.cu(300): error #940-D: missing return statement at end of non-void function "lambda []()->std::tuple<std::variant<cub::CUB_300200_SM_750::detail::transform::RuntimeTransformAgentPrefetchPolicy, cub::CUB_300200_SM_750::detail::transform::RuntimeTransformAgentVectorizedPolicy, cub::CUB_300200_SM_750::detail::transform::RuntimeTransformAgentAsyncPolicy>, std::string>" [T:\src\cccl\python\cuda_cccl\build\cp312-cp312-win_amd64\_parent_cccl\c\parallel\cccl.c.parallel.vcxproj]
          }();

* Fix MSVC size_t/int error in c/parallel/src/nvrtc/command_list.h.

* Fix MSVC size_t/int errors in transform.cu.

* Fix MSVC size_t/int error in cub/cub/detail/launcher/cuda_driver.cuh.

The `CUlaunchConfig` struct and the `CUlaunchKernel` function both take
an unsigned int for the shared memory parameter.

* Suppress the `unreachable code` warnings on Windows in cub/cub/device/dispatch/dispatch_transform.cuh.

* Fix MSVC size_t/int error in c/parallel/src/for.cu.

This particular fix is a bit more thorough than just schlepping in a
static cast; make sure we compute the intermediate sizes in 64-bit and
check the final version for overflow before downcasting to 32-bit.

* Fix MSVC warning C4458: `declaration of 'name' hides class member` in c/parallel/src/jit_templates/templates/operation.h.

* Fix MSVC error `#20013-D: calling a constexpr __host__ ... from __host__ __device__ ...`.

MSVC's C++20 support marks std::string constructors as constexpr, which
is the root of this problem, I believe.  Linux/libstdc++ doesn't have
this issue (yet?).

Simplest solution I could find that actually compiled on Windows is just
to no-op all of the problematic code when doing device compilation w/
nvcc.

* Fix MSVC size_t/int error in c/parallel/src/histogram.cu.

* Fix MSVC size_t/int errors in cub/cub/agent/single_pass_scan_operators.cuh.

For this one I bumped the underlying struct fields to `size_t` as that
felt like a cleaner solution.

* Fix MSVC size_t/int errors in c/parallel's merge_sort.cu, radix_sort.cu, unique_by_key.cu.

These fixes are all similar so I've lumped them together.

* Fix shadowing variable

* Fix MSVC breaking on variadic arguments

* Move warning suppression out of class

This sometimes breaks nvc++

* More Warning suppression

* Fix MSBVC std::format_string-type errors in c/parallel/src/jit_templates/templates/operation.h.

* On Windows, link to DbgHelp for UnDecorateSymbolName.

* Fix MSVC error `'void *': unknown size` in _bindings_impl.c.

MSVC doesn't like pointer arithmetic being done on `void *` addresses.
For the `_get_cubin()` implementations, just explicitly call the
appropriate Python API with a `const char *` cast instead of relying on
array slicing.

* Fix cccl.c.parallel.experimental._bindings_impl import on Windows.

* Avoid a permission denied error on Windows by deferring an os.unlink() call.

The code as written hits a permission denied error on Windows in the finally
block's `os.unlink()`, presumably because the `nvdisasm` process handle
doesn't get GC'd until the function returns.

* Set the RUNTIME_OUTPUT_DIRECTORY for c.parallel.

Without this change, on Windows, the binaries get written to
`d:\bin\Release`.

* Implement helper cccl_type_name_from_nvrtc() to wrap nvrtcGetTypeName().

This is needed to handle post-processing of certain quirky type names on
Windows into a format that nvrtc won't balk on during compilation.

* Bump to numba-cuda 0.20 to fix nvvm issue on Windows.

* Fix cuda.cooperative on Windows.

On Windows, a NamedTemporaryFile can't be re-opened elsewhere (i.e.
numba-cuda) if it hasn't been created with `delete=False`.  With this
fix, cuda.cooperative tests now pass on Windows.

* PR feedback: use cuda::ceil_div() for calculating block size.

Co-authored-by: Michael Schellenberger Costa <miscco@nvidia.com>

* PR feedback: use cuda::ceil_div() for calculating scale.

Co-authored-by: Michael Schellenberger Costa <miscco@nvidia.com>

* PR feedback: const-cast num_options.

Co-authored-by: Bernhard Manfred Gruber <bernhardmgruber@gmail.com>

* PR feedback: const-cast and inline variable decl.

Co-authored-by: Bernhard Manfred Gruber <bernhardmgruber@gmail.com>

* PR feedback: remove unnecessary bounds check.

* PR feedback: use ptrdiff_t for OffsetT for c/parallel/src/transform.cu.

* PR feedback: rework size_t/int fixes for radix_sort.cu, merge_sort.cu, and unique_by_key.cu.

---------

Co-authored-by: Michael Schellenberger Costa <miscco@nvidia.com>
Co-authored-by: Bernhard Manfred Gruber <bernhardmgruber@gmail.com>