Add a 'pull_request_lite' workflow for unmodified dependees.
When a project is actually modified, that project's jobs in the `pull_request` workflow will run.

When a project is not modified but has a modified dependency, that project's jobs in the 'pull-request_lite` matrix will run.

If core CCCL infrastructure is modified, the full `pull_request` workflow will continue to run.