Addressed review feedback
Reimplemented tuning for segmented scan based on
excelled review suggestion to set AgentSegmentedScanPolicyT
template values based on AgentScanPolicyT instance extracted
from the same node in chain of policies.

The assumption is that the chain detail::segmented_scan::policy_hub
is the same as the chain in detail::scan::policy_hub.