Refactor our execution policies
We currently tag our execution policies with just an enumeration for that represents the  standard execution policies.

However, this is not sufficient for our use cases, because we also want to pass along the execution backend and the memory_direction of the algorithm.

This changes our policies so that they take an unsigned integer instead of an enumeration and then adds facilities to set and get the respective properties