Publicly expose `<cuda/std/algorithm>` (#3741)
This ports the remaining standard serial algorithms and publicly exposes the `<cuda/std/algorithm>` header

All algorithms you say. Not quite, we prepared the sorting algorithms, but the recursive implementation commonly used on the CPU is not well suited for GPU execution. Because of that we did not yet expose the sorting algorithms until we reimplement them with an efficient GPU friendly algorithm